pyproject.toml
[project]
name = "convoviz"
version = "0.2.0"
description = "Get analytics and visualizations on your ChatGPT data!"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Mohamed Cheikh Sidiya", email = "mohamedcheikhsidiya77@gmail.com" }
]
dependencies = [
    "matplotlib>=3.9.4",
    "nltk>=3.9.2",
    "orjson>=3.11.5",
    "pillow>=11.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.7.0",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "tqdm>=4.67.1",
    "typer>=0.21.0",
    "wordcloud>=1.9.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["convoviz"]

[project.scripts]
convoviz = "convoviz.cli:main_entry"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "ruff>=0.14.10",
    "types-tqdm>=4.67.0.20250809",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["convoviz"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
